refactor: put the CPU-frequency helpers where their dependency lives
They sat in utils, where anything could import them and then fail at the call on an install without psutil. They belong with the flops suite, the only caller that needs a clock rate, so the gate that covers psutil covers them too. The metadata collector reads them inside the collecting method, as it already does for cpuinfo and psutil, so models keeps no import into a gated package.

The fallback clock constant is renamed to agree with the parameter it defaults, and 'nominal stand-in' is gone from the prose it appeared in.
