+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- the float-vs-`CountedFloat` benchmark no longer prints a cycles column that silently reported nanoseconds whenever the CPU frequency could not be read
+- evaluating counting overhead no longer prints a cycles column that silently reported nanoseconds whenever the CPU frequency could not be read
 
 ### Security
