+++ src/counted_float/_core/benchmarking/flops/_cpu_freq.py
