+++ src/counted_float/_core/utils/__init__.py
@@ -1,7 +1,6 @@
-from ._cpu_freq import get_cpu_frequency_mhz_current, get_cpu_frequency_mhz_max, get_cpu_frequency_mhz_min
 from ._formatting import format_latency, format_time_duration
 from ._geo_mean import geo_mean
-from ._latency import NOMINAL_CPU_FREQ_MHZ, convert_nsecs_to_cycles
+from ._latency import FALLBACK_CPU_FREQ_MHZ, convert_nsecs_to_cycles
 from ._missing_data import impute_missing_data
 from ._quantile import quantile
 from ._rounding import round_number
