+++ tests/benchmarking/micro/test_cycles_are_caller_supplied.py
@@ -6,9 +6,11 @@ reports absolute per-op cost supplies a clock, and one that reports a ratio does
 
 import pytest
 
-import counted_float._core.utils._cpu_freq as cpu_freq_module
 from counted_float._core.benchmarking.micro import InterleavedBenchmarkRunner, MicroBenchmark
 from counted_float._core.models import MicroBenchmarkResult, SingleRunResult
+
+# patched to prove the timing layer never reaches it; it lives behind the benchmarking extra
+cpu_freq_module = pytest.importorskip("counted_float._core.benchmarking.flops._cpu_freq")
 
 
 class _TrivialBenchmark(MicroBenchmark):
